More fixes

+++ codedeploy/scripts/stop.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-systemctl stop spring-pet-clinic.service
+if systemctl is-active --quiet spring-pet-clinic.service; then
+  systemctl stop spring-pet-clinic.service
+fi
